make the dialog component to confirm user deletion

+++ src/app/components/members/members.component.ts
@@ -5,6 +5,7 @@ import { UserService } from 'src/app/api/user.service';
 import { first } from 'rxjs/operators';
 import { DialogService } from 'src/app/dialog/dialog.service';
 import { CreateUserComponent } from './create-user/create-user.component';
+import { PromptDeleteUserComponent } from './prompt-delete-user/prompt-delete-user.component';
 
 @Component({
   selector: 'app-members',
@@ -20,7 +21,7 @@ export class MembersComponent implements OnInit, OnDestroy {
   displayedColumns: string[] = ['name', 'email', 'status', 'action']
   private usersSub: Subscription
 
-  constructor(private userService: UserService, private dialog: DialogService) { 
+  constructor(private userService: UserService, private dialog: DialogService) {
 
     if (this.userService.busy) {
       this.userService.getUsers().pipe(first()).subscribe(this.initMembers)
@@ -30,7 +31,7 @@ export class MembersComponent implements OnInit, OnDestroy {
   }
 
   ngOnInit(): void {
-    
+
   }
 
   private initMembers = (users) => {
@@ -44,16 +45,24 @@ export class MembersComponent implements OnInit, OnDestroy {
     }
   }
   deleteUser(user: User) {
-    this.userService.deleteUser(user)
+    this.dialog.open(PromptDeleteUserComponent, {
+      disableClose: true,
+      closeOnNavigation: false,
+      autoFocus: true
+    })
+      .afterClosed().pipe(first()).subscribe(value => {
+        if (value === true)
+          this.userService.deleteUser(user)
+      })
   }
   createUser() {
     this.dialog.open(CreateUserComponent, {
       disableClose: true,
       closeOnNavigation: false,
       autoFocus: true
-    }).afterClosed().subscribe(user=>{
+    }).afterClosed().subscribe(user => {
       location.reload()
     })
   }
-  
+
 }
